// packages/ocpi-base/src/model/version-number.ts
// SPDX-FileCopyrightText: 2025 Contributors to the CitrineOS Project
//
// SPDX-License-Identifier: Apache-2.0

export enum VersionNumber {
  TWO_DOT_ZERO = '2.0',
  TWO_DOT_ONE = '2.1',
  TWO_DOT_ONE_DOT_ONE = '2.1.1',
  TWO_DOT_TWO = '2.2',
  TWO_DOT_TWO_DOT_ONE = '2.2.1',
}

export const VersionNumberEnumName = 'VersionNumber';